+++ ui/src/main/java/io/snabble/sdk/ui/checkout/CheckoutOnlineView.java
@@ -214,5 +214,15 @@ public class CheckoutOnlineView extends FrameLayout implements Checkout.OnChecko
             helperText.setVisibility(View.VISIBLE);
             progressIndicator.setVisibility(View.GONE);
         }
+
+        if (currentState == Checkout.State.PAYMENT_PROCESSING) {
+            checkoutIdCode.setVisibility(View.GONE);
+            helperText.setVisibility(View.GONE);
+            helperImage.setVisibility(View.GONE);
+            upArrow.setVisibility(View.GONE);
+            progressIndicator.setVisibility(View.VISIBLE);
+            cancel.setVisibility(View.INVISIBLE);
+            cancelProgress.setVisibility(View.INVISIBLE);
+        }
     }
 }